% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuditTools.R
\docType{package}
\name{AuditTools}
\alias{AuditTools}
\alias{AuditTools-package}
\title{AuditTools: tools for analytic testing}
\description{
The AuditTools package gives you the tools to perform analytic
testing and reduce the amount of fieldwork.
}

