% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benford.R
\name{Benford}
\alias{Benford}
\title{Create a Benford distribution}
\usage{
Benford(dat)
}
\arguments{
\item{dat}{vector of}
}
\value{
a matrix with the theoretical and actual distribution
}
\description{
reurns a dataframe of theoretical and actual distributions
}

