% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassPairs.R
\name{ClassPairs}
\alias{ClassPairs}
\title{Classifies the pairs to create a footprint}
\usage{
ClassPairs(L, traces)
}
\arguments{
\item{L}{dataframe from the MakePairs function}

\item{traces}{list of traces}
}
\description{
Classifies each of the pairs as required to draw the Petri net
and the footprint. Pairs are identified by rows in the dataframe,
The first column is the preceding point and the second column is
the following point. There are 4 values the pair can take:
\enumerate{
\item R or Right.  The first value always precedes the second value.
It is called Right because the process flows towards the right.
\item L or Left. The first value always follows the second value. Similarly,
the process flows to the left.
\item # or unrelated. These two values never directly interact. One never directly follows the other.
\item || or parallel.  These two values are do not have an order between them and can happen in either order e.g. AB or BA
}
}

