% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{DataPrep}
\alias{DataPrep}
\title{Convert a dataframe to a list of traces}
\usage{
DataPrep(dat)
}
\arguments{
\item{dat}{2 column dataframe.  The first colum is the key and the second is the event. The data should be sorted by in chronilogical order by event.}
}
\value{
list of traces
}
\description{
Convert a dataframe to a list of traces
}

