% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventsToAlpha.R
\name{EventsToAlpha}
\alias{EventsToAlpha}
\title{Converts dataframes to alphabetical character symbol}
\usage{
EventsToAlpha(x)
}
\arguments{
\item{x}{numeric to be converted to alpha}
}
\value{
alpha
}
\description{
This function converts numeric factor levels to an alpha
characters.  As an example, the first factor level, 1,
would be converted to 'A'.
}

