% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPlaces.R
\name{FindPlaces}
\alias{FindPlaces}
\title{Find the places between the events}
\usage{
FindPlaces(P)
}
\arguments{
\item{P}{dataframe of interactions with flow indicators (#,||,R,L)}
}
\value{
dataframe
}
\description{
Find the places between the events
}

