% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePairs.R
\name{MakePairs}
\alias{MakePairs}
\title{Create pairs for factors that interact in the traces}
\usage{
MakePairs(traces)
}
\arguments{
\item{traces}{list of unique traces}
}
\description{
Create pairs for factors that interact in the traces
}

