% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessMiner.R
\name{ProcessMiner}
\alias{ProcessMiner}
\title{Creates a petri net based to visualize a process}
\usage{
ProcessMiner(dat)
}
\arguments{
\item{dat}{dataframe with two columns. The first column is
the key and second is the event}
}
\description{
Creates a petri net based to visualize a process
}

