% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowGrid.R
\name{ShowGrid}
\alias{ShowGrid}
\title{Show the foot print for a process flow}
\usage{
ShowGrid(x)
}
\arguments{
\item{x}{list of classed pairs from the ClassPairs function}
}
\value{
table
}
\description{
Converts a paired list of relations into a tabular foot print
}

