% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatSample.R
\name{StatSample}
\alias{StatSample}
\title{Create a Statistical sampling object}
\usage{
StatSample(x, method = "MUS", conf = 0.95, target)
}
\arguments{
\item{x}{Dataframe representing the sample population}

\item{method}{Sample selection method.  MUS for Monetary
Unit Sampling and PUS for Physical Object Sampling.}

\item{conf}{Confidence level. Numeric value between 0 an 1 
which will impact the range of the confidence interval.  
Higher confidence levels result in greater confidence interval
range.}

\item{target}{Target of the sampling test. Can be a character 
value matching the column name or the column index value. For 
MUS, this is the column with the numerical data. As an 
example, the transaction dollar amount. If not defined, the
function Will default to the last column of the dataframe.}
}
\description{
This function creates a statistical samplingobject. 
The output will supply sample selections for detailed 
testing. Once testing is complete, the testin 
conclusions can be passed to the predict function to 
return the results of the audit test.
}

